build: fix linux warnings and standardize cmake presets
- Mark vendor includes (FastNoiseLite) as SYSTEM to suppress strict compiler warnings on Linux.
- Fix 'int' to 'float' implicit conversion warning in InputHandler.cpp.
- Add 'CMakePresets.json' to define standard Debug (ASan ON) and Release (ASan OFF) configurations.
- Update .gitignore to track presets but exclude legacy VS settings.

+++ src/app/InputHandler.cpp
@@ -167,7 +167,7 @@ Core::Mat4 InputHandler::GetViewProjectionMatrix() {
 void InputHandler::UpdateTitleInfo(GLFWwindow* pWindow) const {
     if (!pWindow)
         return;
-    int iFPS = static_cast<int>(m_iFrameCount / m_fTimer);
+    int iFPS = static_cast<int>(static_cast<float>(m_iFrameCount) / m_fTimer);
     std::string strTitle = "HPC Voxel Engine FPS:" + std::to_string(iFPS);
     if (m_bEnableFaceCulling)
         strTitle += "\tFace Culling Enabled (Press 'F' key to toogle)";
